Z2 - https://android.mpage.sk/cvicenie3.php - navigation between List of pubs and detail of a pub

+++ app/src/main/java/com/example/semestralne_zadanie/model/Pub.kt
@@ -1,4 +1,4 @@
 package com.example.semestralne_zadanie.model
 
-class Pub (val type: String, val id: Int, val lat: Double, val alt: Double, val tags: Map<String, String>) {
+class Pub (val type: String, val id: Double, val lat: Double, val alt: Double, val tags: Map<String, String>) {
 }
